+++ docs/source/conf.py
@@ -84,10 +84,14 @@ html_favicon = '_static/favicon.ico'
 
 # -- Extension configuration -------------------------------------------------
 
-# Napoleon settings
+# Napoleon settings (Google-style docstrings)
+napoleon_google_docstring = True
+napoleon_numpy_docstring = False
 napoleon_include_init_with_doc = True
 napoleon_include_private_with_doc = True
 napoleon_use_admonition_for_examples = True
+napoleon_use_admonition_for_notes = True
+napoleon_use_admonition_for_references = True
 
 # autosummary settings
 autodoc_member_order = 'bysource'